Move epochs begin block code to keeper (#1870)
